requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
alembic==1.13.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.9
